backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
Pillow>=10.0
reportlab>=4.0
requests>=2.31.0
python-decouple>=3.8
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.0.0
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.5.0
psycopg2-binary>=2.9.9
django-redis>=6.0.0
django-filter>=23.5
whitenoise>=6.6.0
gunicorn>=21.2.0
sentry-sdk[django]>=1.40.0
django-axes>=6.4.0
cloudinary>=1.38.0
django-cloudinary-storage>=0.3.0
